// src/types.ts
export type Language = 'en' | 'ps' | 'dr';